Minor fixes and cleanup.
Fixed incorrect errno and casts for implicit conversions that static code
analysis complained about. Added explanation for infinite planes for
dsFrustum3*.

Updated errno section of core README.md. Also added an explanation as to
why errno is used, not just to explain to others but also to remind myself
whenever I think that errno feels bad to use. (yes it is bad, but the added
API and usage complexity is arguably worse given that it's only for logging
info)

+++ modules/Render/Render/src/CommandBuffer.c
@@ -97,8 +97,8 @@ bool dsCommandBuffer_beginSecondary(dsCommandBuffer* commandBuffer,
 	}
 
 	if (viewport && (viewport->min.x < 0 || viewport->min.y < 0 || viewport->min.z < 0 ||
-		viewport->max.z > 1 || (framebuffer && (viewport->max.x > framebuffer->width ||
-			viewport->max.y > framebuffer->height))))
+		viewport->max.z > 1 || (framebuffer && (viewport->max.x > (float)framebuffer->width ||
+			viewport->max.y > (float)framebuffer->height))))
 	{
 		errno = ERANGE;
 		DS_LOG_ERROR(DS_RENDER_LOG_TAG, "Viewport is out of range.");
